refactor(hookengine): extract FocusOwner (Wave 3 PR 3.2)
* New: src/app/system/FocusOwner.{h,cpp} (~640 LOC total)
  - Owns focus state + WinEvent hooks + window classification
  - Public surface: Install/Uninstall, Classify (ConfigContext-based),
    RefreshFocusCache, current/previous exe, smart-switch + CJK layout
    accessors
  - GetExeNameForHwnd remains a free Win32 helper (now static method)
  - WinEventProc dispatches to host via FocusChangedFn callback;
    Chromium/Java reinstall via ReinstallFn (host gates on lifecycle)
* HookEngine: focus_ member alongside lifecycle_ (declared BEFORE so
  destruction order runs ~lifecycle_ first, then ~focus_). Apply*
  handlers read from focus_ via getters; OnLayoutChanged/CheckLayoutChange
  stay here (engine state writes). HookEngine.cpp 4431 → 3868 LOC
  (-563 net; ~600 method bodies deleted, ~37 added for migration comments
  and call-site rewrites).
* HookEngine.h: 14 state members migrated (focusHook_, minimizeHook_,
  lastForegroundPid_, currentExe_/previousExe_, appProfileCache_+struct,
  appModeMap_/appModeDirty_/smartSwitchMgr_, layoutSuppressed_/
  modeBeforeCjk_/cachedIsCompatLayout_, cachedFocusedHwnd_/
  cachedFocusedClass_, webView2PositiveCache_).
* CMake: VKeyApp + VKeyLite include the new translation unit
* Audit: check_hook_thread_no_mutex.sh Check 1 threshold dropped 3 → 2
  (WinEventProc moved out of HookEngine.cpp; FocusOwner has no stateMutex_
  so the trap is moot there). All other checks unchanged.
* Tests: Linux GTest 2011/2011 PASS

Refs docs/plans/2026-05-24-wave3-pr3.2-focusowner-handoff.md.
