+++ src/app/system/FocusOwner.cpp
@@ -1,0 +1,640 @@
+// VKey - Focus Owner Implementation
+// SPDX-License-Identifier: GPL-3.0-only
+
+#include "FocusOwner.h"
+#include "HookLifecycle.h"  // REINSTALL_REASON_CHROMIUM / _JAVA
+#include "PerfHistogram.h"  // PERF_SCOPE
+#include "helpers/AppHelpers.h"  // ToLowerAscii, GetFocusedChildHwnd
+#include "core/CrashLog.h"
+#include "core/Debug.h"
+#include "core/Logger.h"
+
+#include <tlhelp32.h>
+#include <cstdint>
+#include <utility>
+
+namespace NextKey {
+
+// Same prefix-only pattern as HookEngine.cpp / HookLifecycle.cpp: arg side
+// effects are not evaluated when the runtime logger is off.
+#define FOCUS_LOG(fmt, ...) do {                                              \
+    if (::NextKey::Logger::IsEnabled())                                       \
+        ::NextKey::Logger::Log(L"[Hook] " fmt, ##__VA_ARGS__);                \
+} while (0)
+
+std::atomic<FocusOwner*> FocusOwner::s_instance{nullptr};
+
+// ─────────────────────────────────────────────────────────────────────────
+// File-scope helpers (moved from HookEngine.cpp). Kept anonymous-namespace-
+// free for symmetry with the originals and to keep diffs small if a future
+// PR needs to reorder them.
+// ─────────────────────────────────────────────────────────────────────────
+
+/// Check if a filename (without path) is a known Electron app executable.
+/// Electron apps use Chrome_WidgetWin window class (same as Chromium browsers).
+/// Unknown Chrome_WidgetWin apps default to "browser" — safer because:
+///   - Browser miss → double text (visible, user reports immediately)
+///   - Electron miss → slightly slower input (split delay absent, usually OK)
+/// This list covers the most popular Electron apps. Add new ones as needed.
+static bool IsKnownElectronExe(const wchar_t* filename) noexcept {
+    return _wcsnicmp(filename, L"code", 4) == 0 ||       // VS Code
+           _wcsnicmp(filename, L"cursor", 6) == 0 ||     // Cursor (AI code editor)
+           _wcsnicmp(filename, L"discord", 7) == 0 ||    // Discord
+           _wcsnicmp(filename, L"slack", 5) == 0 ||      // Slack
+           _wcsnicmp(filename, L"notion", 6) == 0 ||     // Notion
+           _wcsnicmp(filename, L"obsidian", 8) == 0 ||   // Obsidian
+           _wcsnicmp(filename, L"figma", 5) == 0 ||      // Figma
+           _wcsnicmp(filename, L"postman", 7) == 0 ||    // Postman
+           _wcsnicmp(filename, L"insomnia", 8) == 0 ||   // Insomnia
+           _wcsnicmp(filename, L"signal", 6) == 0 ||     // Signal
+           _wcsnicmp(filename, L"1password", 9) == 0 ||  // 1Password
+           _wcsnicmp(filename, L"bitwarden", 9) == 0 ||  // Bitwarden
+           _wcsnicmp(filename, L"gitkraken", 9) == 0 ||  // GitKraken
+           _wcsnicmp(filename, L"hyper", 5) == 0 ||      // Hyper terminal
+           _wcsnicmp(filename, L"spotify", 7) == 0 ||    // Spotify
+           _wcsnicmp(filename, L"whatsapp", 8) == 0 ||   // WhatsApp Desktop
+           _wcsnicmp(filename, L"telegram", 8) == 0 ||   // Telegram (some forks are Electron; native is Qt, caught earlier)
+           _wcsnicmp(filename, L"logseq", 6) == 0 ||     // Logseq
+           _wcsnicmp(filename, L"linear", 6) == 0 ||     // Linear
+           _wcsnicmp(filename, L"lark", 4) == 0 ||       // Lark/Feishu
+           _wcsnicmp(filename, L"zalo", 4) == 0;         // Zalo PC
+}
+
+/// Returns the full exe path (original case) for the process owning `hwnd`.
+/// Empty on failure. Needed by IsWebView2App() to locate sibling DLLs.
+[[nodiscard]] static std::wstring GetExeFullPathForHwnd(HWND hwnd) noexcept {
+    if (!hwnd) return {};
+    DWORD pid = 0;
+    GetWindowThreadProcessId(hwnd, &pid);
+    if (!pid) return {};
+    HANDLE hProc = OpenProcess(PROCESS_QUERY_LIMITED_INFORMATION, FALSE, pid);
+    if (!hProc) return {};
+    wchar_t exePath[MAX_PATH] = {};
+    DWORD size = MAX_PATH;
+    std::wstring result;
+    if (QueryFullProcessImageNameW(hProc, 0, exePath, &size)) {
+        result.assign(exePath, size);
+    }
+    CloseHandle(hProc);
+    return result;
+}
+
+/// Classify window into app type — called from Classify().
+/// Reads window class ONCE and determines: Browser, Electron, Qt, Console, or Normal.
+/// Results are written directly to the caller's output variables.
+///
+/// Priority order:
+///   1. Console (by window class: ConsoleWindowClass, CASCADIA, mintty, PuTTY)
+///   2. Firefox-based browser (by window class: MozillaWindowClass)
+///   3. Chrome_WidgetWin → Known Electron list or default to browser
+///   4. Qt app (by window class: Qt5*, Qt6*, QWidget)
+///   5. VB6 app (by window class: ThunderRT6*) — needs clipboard paste
+///   6. Normal Win32 app
+static void ClassifyWindow(HWND hwnd,
+                           bool& outIsBrowser,
+                           bool& outIsElectron,
+                           bool& outIsQtApp,
+                           bool& outIsConsole,
+                           bool& outIsVB6) noexcept {
+    outIsBrowser = outIsElectron = outIsQtApp = outIsConsole = outIsVB6 = false;
+
+    HWND root = GetAncestor(hwnd, GA_ROOT);
+    if (root) hwnd = root;
+
+    wchar_t className[64] = {};
+    GetClassNameW(hwnd, className, 64);
+
+    // 1a. Windows Terminal — modern DirectX renderer + ConPTY, handles batch input fine.
+    //     Treated as normal app (no flags set, batch dispatch, no bait).
+    if (_wcsicmp(className, L"CASCADIA_HOSTING_WINDOW_CLASS") == 0) {
+        return;  // No flags set → batch path
+    }
+
+    // 1b. Legacy console apps — outIsConsole triggers split dispatch in DispatchSendInput
+    if (_wcsicmp(className, L"ConsoleWindowClass") == 0 ||
+        _wcsicmp(className, L"tty") == 0 ||                            // Cygwin/MSYS
+        _wcsicmp(className, L"mintty") == 0 ||                         // Git Bash
+        _wcsicmp(className, L"PuTTY") == 0) {
+        outIsConsole = true;
+        return;
+    }
+
+    // 2. Firefox-based browsers (covers Firefox, Floorp, Tor, LibreWolf, Waterfox, Pale Moon)
+    if (_wcsicmp(className, L"MozillaWindowClass") == 0) {
+        outIsBrowser = true;
+        return;
+    }
+
+    // 3. Chrome_WidgetWin: Chromium browser OR Electron app
+    //    Disambiguate by known Electron exe list. Unknown → browser (safer default).
+    if (wcsstr(className, L"Chrome_WidgetWin")) {
+        std::wstring exeName = FocusOwner::GetExeNameForHwnd(hwnd);
+        if (!exeName.empty() && IsKnownElectronExe(exeName.c_str())) {
+            outIsElectron = true;
+        } else {
+            outIsBrowser = true;  // Unknown Chrome_WidgetWin → assume browser
+        }
+        return;
+    }
+
+    // 4. Qt apps (Telegram native, KeePassXC, etc.)
+    if (wcsstr(className, L"Qt5") || wcsstr(className, L"Qt6") ||
+        wcsstr(className, L"QWidget")) {
+        outIsQtApp = true;
+        return;
+    }
+
+    // 5. VB6 apps (XYplorer, etc.): register Unicode window classes but process
+    //    messages as ANSI internally — KEYEVENTF_UNICODE / VK_PACKET chars become '?'.
+    if (_wcsnicmp(className, L"ThunderRT6", 10) == 0) {
+        outIsVB6 = true;
+        return;
+    }
+
+    // 6. Normal Win32 app (Notepad, Word, etc.) — no flags set
+}
+
+// ─────────────────────────────────────────────────────────────────────────
+// FocusOwner — public interface
+// ─────────────────────────────────────────────────────────────────────────
+
+FocusOwner::FocusOwner() = default;
+
+FocusOwner::~FocusOwner() {
+    Uninstall();
+}
+
+bool FocusOwner::Install(FocusChangedFn onFocusChanged,
+                          ReinstallFn   onReinstall) {
+    if (focusHook_) return true;  // already installed — idempotent
+
+    onFocusChanged_ = std::move(onFocusChanged);
+    onReinstall_    = std::move(onReinstall);
+    s_instance.store(this, std::memory_order_release);
+
+    // Two separate hooks for exact event targeting (avoids receiving ~20
+    // unrelated events in the 0x0003..0x0017 range).
+    focusHook_ = SetWinEventHook(
+        EVENT_SYSTEM_FOREGROUND, EVENT_SYSTEM_FOREGROUND,
+        nullptr, WinEventProc,
+        0, 0, WINEVENT_OUTOFCONTEXT | WINEVENT_SKIPOWNPROCESS);
+
+    // MINIMIZEEND: restoring a window from the taskbar may not fire FOREGROUND
+    // (taskbar gets the foreground event, filtered as Shell_TrayWnd).
+    minimizeHook_ = SetWinEventHook(
+        EVENT_SYSTEM_MINIMIZEEND, EVENT_SYSTEM_MINIMIZEEND,
+        nullptr, WinEventProc,
+        0, 0, WINEVENT_OUTOFCONTEXT | WINEVENT_SKIPOWNPROCESS);
+
+    if (!focusHook_) {
+        s_instance.store(nullptr, std::memory_order_release);
+        onFocusChanged_ = nullptr;
+        onReinstall_    = nullptr;
+        if (minimizeHook_) {
+            UnhookWinEvent(minimizeHook_);
+            minimizeHook_ = nullptr;
+        }
+        return false;
+    }
+    return true;
+}
+
+void FocusOwner::Uninstall() {
+    if (focusHook_) {
+        UnhookWinEvent(focusHook_);
+        focusHook_ = nullptr;
+    }
+    if (minimizeHook_) {
+        UnhookWinEvent(minimizeHook_);
+        minimizeHook_ = nullptr;
+    }
+    if (s_instance.load(std::memory_order_acquire) == this) {
+        s_instance.store(nullptr, std::memory_order_release);
+    }
+    onFocusChanged_ = nullptr;
+    onReinstall_    = nullptr;
+}
+
+void CALLBACK FocusOwner::WinEventProc(HWINEVENTHOOK, DWORD event, HWND hwnd,
+                                        LONG, LONG, DWORD, DWORD) {
+    try {
+        FocusOwner* self = s_instance.load(std::memory_order_relaxed);
+        if (!self || !self->onFocusChanged_) return;
+
+        // Per WINEVENT_OUTOFCONTEXT semantics, this callback fires on the
+        // INSTALLER thread (main, where SetWinEventHook was called) — NOT
+        // the hook thread. State writes are deferred to the hook thread via
+        // the mailbox (HookEngine::OnFocusChanged → Classify → Post).
+        if (event == EVENT_SYSTEM_MINIMIZEEND) {
+            // Window restored from taskbar — re-evaluate focus with the actual foreground window.
+            // Don't use hwnd directly: the restored window may not be foreground yet.
+            FOCUS_LOG(L"MINIMIZEEND (hwnd=%p) — re-evaluating focus", hwnd);
+            self->onFocusChanged_(nullptr);  // nullptr → uses GetForegroundWindow()
+            return;
+        }
+
+        FOCUS_LOG(L"FOCUS changed (hwnd=%p) — classifying + posting", hwnd);
+        self->onFocusChanged_(hwnd);
+    } catch (const std::exception& e) {
+        CrashLog(L"FocusOwner::WinEventProc", e.what());
+    } catch (...) {
+        CrashLog(L"FocusOwner::WinEventProc", "(non-std exception)");
+    }
+}
+
+// ─────────────────────────────────────────────────────────────────────────
+// Cache helpers (hook-thread only — same invariant as pre-PR-3.2: only
+// Classify (main/worker) and ApplyFocusOnHookThread (hook) touch the maps.
+// Pre-existing single-thread guarantee carried over verbatim.)
+// ─────────────────────────────────────────────────────────────────────────
+
+const FocusOwner::AppProfile* FocusOwner::LookupAppProfile(HWND hwnd) noexcept {
+    auto it = appProfileCache_.find(hwnd);
+    if (it == appProfileCache_.end()) return nullptr;
+
+    // Validate: HWND values can be reused after the owning process dies.
+    // GetWindowThreadProcessId is one cheap syscall; on hit it still saves
+    // the much pricier ClassifyWindow + GetExeNameForHwnd + IsWebView2App
+    // child-window walk that we'd otherwise rerun.
+    DWORD currentPid = 0;
+    GetWindowThreadProcessId(hwnd, &currentPid);
+    if (currentPid == 0 || currentPid != it->second.pid) {
+        appProfileCache_.erase(it);
+        return nullptr;
+    }
+    return &it->second;
+}
+
+void FocusOwner::StoreAppProfile(HWND hwnd, AppProfile profile) noexcept {
+    profile.cachedAt = GetTickCount64();
+
+    // Bounded cache: LRU-evict the oldest entry when at capacity. O(N) scan
+    // is fine — N is capped at kMaxAppProfileCache (64).
+    if (appProfileCache_.size() >= kMaxAppProfileCache) {
+        auto oldest = appProfileCache_.begin();
+        for (auto it = std::next(appProfileCache_.begin());
+             it != appProfileCache_.end(); ++it) {
+            if (it->second.cachedAt < oldest->second.cachedAt) oldest = it;
+        }
+        appProfileCache_.erase(oldest);
+    }
+
+    appProfileCache_[hwnd] = std::move(profile);
+}
+
+bool FocusOwner::IsTrayOrTaskbarWindow(HWND hwnd) noexcept {
+    if (!hwnd) return false;
+
+    // Ignore focus switches to our own process (Settings, Menu, Tray)
+    DWORD processId;
+    GetWindowThreadProcessId(hwnd, &processId);
+    if (processId == GetCurrentProcessId()) {
+        return true;
+    }
+
+    // GetAncestor is a no-op when hwnd is already a root (e.g. from GetForegroundWindow),
+    // but needed when called with a child HWND (e.g. from WindowFromPoint).
+    HWND root = GetAncestor(hwnd, GA_ROOT);
+    if (root) hwnd = root;
+    wchar_t cls[64] = {};
+    GetClassNameW(hwnd, cls, 64);
+    return _wcsicmp(cls, L"Shell_TrayWnd") == 0 ||            // main taskbar
+           _wcsicmp(cls, L"TrayNotifyWnd") == 0 ||            // notification area
+           _wcsicmp(cls, L"NotifyIconOverflowWindow") == 0 ||  // overflow (^) Win 10
+           _wcsicmp(cls, L"TopLevelWindowForOverflowTray") == 0 || // overflow (^) Win 11
+           _wcsicmp(cls, L"Shell_SecondaryTrayWnd") == 0 ||   // secondary taskbar
+           _wcsicmp(cls, L"XamlExplorerHostIslandWindow") == 0 || // Win 11 tray popups (volume, network)
+           _wcsicmp(cls, L"#32768") == 0 ||                   // standard popup menu (right-click tray apps)
+           _wcsicmp(cls, L"MSTaskSwWClass") == 0 ||            // taskbar app buttons
+           _wcsicmp(cls, L"Start") == 0 ||                     // Start button
+           _wcsicmp(cls, L"Windows.UI.Core.CoreWindow") == 0 || // Start Menu / Action Center (Win 10/11)
+           _wcsicmp(cls, L"VKeyTrayClass") == 0;           // VKey own tray window
+           // Note: SetForegroundWindow(hwndMessage_) in ShowContextMenu fires
+           // EVENT_SYSTEM_FOREGROUND synchronously, but WinEventProc is WINEVENT_OUTOFCONTEXT
+           // so it's delivered asynchronously — this filter still catches it correctly.
+}
+
+std::wstring FocusOwner::GetExeNameForHwnd(HWND hwnd) noexcept {
+    if (!hwnd) return {};
+    DWORD pid = 0;
+    GetWindowThreadProcessId(hwnd, &pid);
+    if (!pid) return {};
+    HANDLE hProc = OpenProcess(PROCESS_QUERY_LIMITED_INFORMATION, FALSE, pid);
+    if (!hProc) return {};
+    wchar_t exePath[MAX_PATH] = {};
+    DWORD size = MAX_PATH;
+    std::wstring result;
+    if (QueryFullProcessImageNameW(hProc, 0, exePath, &size)) {
+        const wchar_t* filename = wcsrchr(exePath, L'\\');
+        result = ToLowerAscii(filename ? filename + 1 : exePath);
+    }
+    CloseHandle(hProc);
+    return result;
+}
+
+// ── Auto-detect WebView2 apps via process inspection ──────────────────
+// Tauri apps (Dorion), Office WebView2 add-ins, and any Win32 app that
+// hosts a WebView2 control needs the Electron input treatment (skip
+// reinjectVk + split dispatch) — Chromium's untrusted-input filter drops
+// the unpaired synthetic VK keydown under some conditions.
+//
+// Two-pass detection:
+//   1. Module check: Win32 apps that load WebView2 inline into the host
+//      process will have `WebView2Loader.dll` or `embeddedbrowserwebview.dll`
+//      loaded. Cheap (~3ms) and catches most hybrid Win32 apps.
+//   2. Child-process check: Tauri v2 + modern WebView2 runtime isolate
+//      the browser into `msedgewebview2.exe` — spawned as a child of the
+//      host. The host itself may not load any WebView2 DLL. Walk the
+//      process table and look for a child with that exe name. Slower
+//      (~5-10ms over ~200 processes), so we only run it as a fallback.
+//
+// Cache strategy: positive-only. Both checks race with WebView2 runtime
+// initialization (Tauri delay-loads on first embed), so a false at app-
+// launch time must not poison subsequent checks.
+//
+// Threading: `webView2PositiveCache_` is plain (not atomic). Called only
+// from Classify, which runs on either main (WinEventProc) or worker
+// (OnTickPoll) thread — never both simultaneously (both go through
+// HookEngine::OnFocusChanged which serializes via Windows message loop).
+bool FocusOwner::IsWebView2App(HWND topLevel,
+                                const std::wstring& exeFullPath) noexcept {
+    if (!topLevel || exeFullPath.empty()) return false;
+
+    if (webView2PositiveCache_.count(exeFullPath)) return true;
+
+    DWORD pid = 0;
+    GetWindowThreadProcessId(topLevel, &pid);
+    if (!pid) return false;
+
+    // Instrumentation: snapshot APIs below cross process boundaries (loader lock +
+    // potential AV hook). Logged so 1-off user reports of post-unlock CPU spikes
+    // can be triaged with evidence instead of speculation. See docs/TODO.md
+    // "IsWebView2App perf instrumentation".
+    const ULONGLONG t0 = GetTickCount64();
+    const wchar_t* slash = wcsrchr(exeFullPath.c_str(), L'\\');
+    const wchar_t* exeBase = slash ? slash + 1 : exeFullPath.c_str();
+    const wchar_t* pass1Result = L"snap_fail";
+    const wchar_t* pass2Result = L"skip";
+
+    bool found = false;
+
+    // Pass 1 — loaded modules in the host process.
+    // TH32CS_SNAPMODULE | TH32CS_SNAPMODULE32 covers both native + WoW64.
+    // INVALID_HANDLE_VALUE on cross-IL / AppContainer targets → fall through.
+    if (HANDLE modSnap = CreateToolhelp32Snapshot(TH32CS_SNAPMODULE | TH32CS_SNAPMODULE32, pid);
+        modSnap != INVALID_HANDLE_VALUE) {
+        MODULEENTRY32W me = { sizeof(me) };
+        for (BOOL ok = Module32FirstW(modSnap, &me); ok; ok = Module32NextW(modSnap, &me)) {
+            if (_wcsicmp(me.szModule, L"WebView2Loader.dll") == 0 ||
+                _wcsicmp(me.szModule, L"embeddedbrowserwebview.dll") == 0) {
+                found = true;
+                break;
+            }
+        }
+        CloseHandle(modSnap);
+        pass1Result = found ? L"module_found" : L"module_notfound";
+    }
+
+    // Pass 2 — `msedgewebview2.exe` spawned as a child process.
+    // Covers modern Tauri where the host doesn't load WebView2 DLLs itself.
+    if (!found) {
+        pass2Result = L"snap_fail";
+        if (HANDLE procSnap = CreateToolhelp32Snapshot(TH32CS_SNAPPROCESS, 0);
+            procSnap != INVALID_HANDLE_VALUE) {
+            PROCESSENTRY32W pe = { sizeof(pe) };
+            for (BOOL ok = Process32FirstW(procSnap, &pe); ok; ok = Process32NextW(procSnap, &pe)) {
+                if (pe.th32ParentProcessID == pid &&
+                    _wcsicmp(pe.szExeFile, L"msedgewebview2.exe") == 0) {
+                    found = true;
+                    break;
+                }
+            }
+            CloseHandle(procSnap);
+            pass2Result = found ? L"proc_found" : L"proc_notfound";
+        }
+    }
+
+    if (found) webView2PositiveCache_.insert(exeFullPath);
+
+    FOCUS_LOG(L"  IsWebView2App: pid=%u exe=\"%s\" pass1=%s pass2=%s result=%d dur=%llums",
+             pid, exeBase, pass1Result, pass2Result, found ? 1 : 0,
+             GetTickCount64() - t0);
+    return found;
+}
+
+void FocusOwner::RefreshFocusCache(HWND foreground) noexcept {
+    HWND focused = ::NextKey::GetFocusedChildHwnd(foreground);
+    cachedFocusedHwnd_.store(focused, std::memory_order_relaxed);
+    if (focused) {
+        wchar_t cls[64] = {};
+        GetClassNameW(focused, cls, 64);
+        cachedFocusedClass_.assign(cls);
+    } else {
+        cachedFocusedClass_.clear();
+    }
+}
+
+void FocusOwner::InvalidateFocusCache() noexcept {
+    cachedFocusedHwnd_.store(nullptr, std::memory_order_relaxed);
+    cachedFocusedClass_.clear();  // tuple race benign — see header doc
+}
+
+// ─────────────────────────────────────────────────────────────────────────
+// Phase 2b classify (moved from HookEngine::ClassifyFocusedWindow).
+//
+// Caller thread: main (WinEventProc) or worker (OnTickPoll). Heavy Win32
+// inspection lives here: ClassifyWindow + GetExeNameForHwnd +
+// IsWebView2App + cache lookup/store + override-map reads. None of these
+// are safe to run from the LL hook callback (Rule 11.2:
+// CreateToolhelp32Snapshot ≥ 3 ms blows the 30 ms p99 Tier-2 budget).
+//
+// The result is a `FocusClassification` POD which HookEngine posts to the
+// mailbox; the hook thread consumes it in ApplyFocusOnHookThread.
+// ─────────────────────────────────────────────────────────────────────────
+FocusClassification FocusOwner::Classify(HWND triggerHwnd,
+                                          const ConfigContext& ctx) noexcept {
+    PERF_SCOPE(::NextKey::Perf::Stage::FocusClassify);
+    FocusClassification cls;
+
+    HWND fg = GetForegroundWindow();
+    // Prefer triggerHwnd (captured at WinEventProc event time): GetForegroundWindow
+    // is async-stale by the time WINEVENT_OUTOFCONTEXT dispatches, often returning
+    // a transient JumpList / taskbar HWND instead of the app the user switched to.
+    HWND activeHwnd = triggerHwnd ? triggerHwnd : fg;
+    if (!activeHwnd) return cls;  // empty cls = "nothing to apply" sentinel
+
+    cls.hwndOpaque = reinterpret_cast<std::uintptr_t>(activeHwnd);
+
+    // Hidden helpers, tray, zero-size, tool windows — still classify (so
+    // dispatch flags stay consistent when focus transits through one) but
+    // skip the smart-switch / currentExe_ update.
+    if (!IsWindowVisible(activeHwnd) || IsIconic(activeHwnd) || IsTrayOrTaskbarWindow(activeHwnd)) {
+        cls.skipAppTracking = true;
+    } else {
+        RECT rect;
+        if (GetWindowRect(activeHwnd, &rect) &&
+            (rect.right - rect.left <= 0 || rect.bottom - rect.top <= 0 || rect.left <= -20000)) {
+            cls.skipAppTracking = true;  // trick message-pump windows (IDM et al.)
+        } else if (GetWindowLongW(activeHwnd, GWL_EXSTYLE) & WS_EX_TOOLWINDOW) {
+            cls.skipAppTracking = true;  // tooltips, context menus, floating helpers
+        }
+    }
+
+    // Cache lookup: skip ClassifyWindow + GetExeNameForHwnd + IsWebView2App
+    // when (HWND, PID) already classified. PID re-check inside LookupAppProfile
+    // catches HWND reuse after a process dies.
+    const AppProfile* cached = LookupAppProfile(activeHwnd);
+    bool isBrowser{}, isElectron{}, isQtApp{}, isVB6{}, localConsole{}, isWebView2{};
+    if (cached) {
+        isBrowser    = cached->isBrowser;
+        isElectron   = cached->isElectron;
+        isQtApp      = cached->isQtApp;
+        localConsole = cached->isConsole;
+        isVB6        = cached->isVB6;
+        isWebView2   = cached->isWebView2;
+        cls.exeName  = cached->exeName;
+    } else {
+        ClassifyWindow(activeHwnd, isBrowser, isElectron, isQtApp, localConsole, isVB6);
+        cls.exeName = GetExeNameForHwnd(activeHwnd);
+    }
+    cls.isBrowser  = isBrowser;
+    cls.isElectron = isElectron;
+    cls.isQtApp    = isQtApp;
+    cls.isVB6      = isVB6;
+    cls.isConsole  = localConsole;
+
+    // Per-app send-method override (clipboard injector toggle). Reads
+    // the snapshot's `appSendMethodOverrides` map — RCU-published from
+    // the worker thread, so this main-thread lookup is lock-free and
+    // immune to torn reads during a TOML rebuild.
+    if (!cls.exeName.empty() && ctx.snap) {
+        auto it = ctx.snap->appSendMethodOverrides.find(cls.exeName);
+        if (it != ctx.snap->appSendMethodOverrides.end() && it->second == 1) {
+            cls.localUseClipboardInjector = true;
+        }
+    }
+
+    // Dispatch-shape derivation. See OnFocusChanged's pre-Phase-2b comments
+    // for the per-host rationale (bait char, split dispatch, edit message
+    // path, clipboard fallback). Preserved verbatim — this is data the
+    // hook hot path will read via the IOutputInjector picked in Apply.
+    cls.localSkipEmpty = cls.isElectron || cls.isConsole;
+    cls.localNeedBait  = cls.isBrowser;
+    cls.localClipboard = cls.isVB6;
+    if (!cls.localSkipEmpty && !cls.localNeedBait && !cls.localClipboard) {
+        if (!cls.exeName.empty()) {
+            if (_wcsicmp(cls.exeName.c_str(), L"zed.exe") == 0) {
+                cls.localSkipEmpty = true;
+            } else if (_wcsicmp(cls.exeName.c_str(), L"notepad.exe") == 0) {
+                // Win11 WinUI 3 Notepad: RichEditBox async on compositor; EM_REPLACESEL
+                // on the child Edit is atomic and avoids the flicker that SendInput
+                // batching causes here. Classic Notepad benefits too (single undo entry).
+                cls.localEditMsg = true;
+            } else {
+                const bool isOutlook = cls.exeName.find(L"outlook") != std::wstring::npos;
+                cls.localNeedBait = cls.exeName.find(L"excel") != std::wstring::npos || isOutlook;
+                if (!cls.localNeedBait) {
+                    if (!cached) {
+                        std::wstring exeFullPath = GetExeFullPathForHwnd(activeHwnd);
+                        isWebView2 = IsWebView2App(activeHwnd, exeFullPath);
+                    }
+                    if (isWebView2) {
+                        cls.localNeedBait = true;
+                        cls.localSkipEmpty = false;
+                    }
+                }
+            }
+        }
+    }
+    cls.isWebView2 = isWebView2;
+    cls.localElectronApp = (cls.isElectron || cls.isWebView2) && !cls.isConsole;
+
+    // Cache miss path: persist for next focus event. Skipped when
+    // GetWindowThreadProcessId fails — invariant requires PID for re-check.
+    DWORD pid = 0;
+    GetWindowThreadProcessId(activeHwnd, &pid);
+    cls.pid = static_cast<std::uint32_t>(pid);
+    if (!cached && pid != 0) {
+        AppProfile profile;
+        profile.pid = pid;
+        profile.exeName = cls.exeName;
+        profile.isBrowser  = isBrowser;
+        profile.isElectron = isElectron;
+        profile.isQtApp    = isQtApp;
+        profile.isConsole  = localConsole;
+        profile.isVB6      = isVB6;
+        profile.isWebView2 = isWebView2;
+        StoreAppProfile(activeHwnd, std::move(profile));
+    }
+
+    // exeName fallback: triggerHwnd may have died by the time the async
+    // event dispatches; fall back to current foreground.
+    if (cls.exeName.empty() && activeHwnd != fg && fg) {
+        cls.exeName = GetExeNameForHwnd(fg);
+    }
+
+    // Java detection — used to trigger top-of-chain hook reinstall (jnativehook
+    // GC stalls regularly exceed LowLevelHooksTimeout and Windows drops us).
+    cls.isJavaApp =
+        cls.exeName == L"jp2launcher.exe" ||
+        cls.exeName == L"javaw.exe" ||
+        cls.exeName == L"java.exe";
+
+    // Per-app excluded / TSF / encoding / method overrides — captured here
+    // so ApplyFocusOnHookThread doesn't need to touch the maps OR read
+    // `global{CodeTable,InputMethod}_` (both written from main; the new
+    // cross-thread read would be a race). Phase 3 RCU-snapshots the maps
+    // and the global values come in via ConfigContext (HookEngine reads
+    // the atomics on the caller thread and hands the resolved values in).
+    cls.targetCodeTable = ctx.globalCodeTable;
+    cls.targetMethod    = ctx.globalInputMethod;
+    if (!cls.skipAppTracking && !cls.exeName.empty() && ctx.snap && ctx.cfg) {
+        if (ctx.cfg->excludeApps && !ctx.snap->excludedAppSet.empty()) {
+            cls.isExcluded = ctx.snap->excludedAppSet.count(cls.exeName) > 0;
+        }
+        if (!cls.isExcluded && ctx.cfg->tsfApps && !ctx.snap->tsfAppSet.empty()) {
+            cls.isTsf = ctx.snap->tsfAppSet.count(cls.exeName) > 0;
+        }
+        if (!cls.isExcluded && !cls.isTsf) {
+            auto itEnc = ctx.snap->appEncodingOverrides.find(cls.exeName);
+            if (itEnc != ctx.snap->appEncodingOverrides.end()) {
+                cls.targetCodeTable = static_cast<int>(itEnc->second);
+            }
+            auto itIm = ctx.snap->appInputMethodOverrides.find(cls.exeName);
+            if (itIm != ctx.snap->appInputMethodOverrides.end()) {
+                cls.targetMethod = static_cast<int>(itIm->second);
+            }
+        }
+    }
+
+    // Re-install hooks to guarantee VKey remains at the top of the hook chain.
+    // Two distinct triggers, two distinct mechanisms:
+    //   1. Chromium-based (Electron, WebView2, Browsers): they install their own
+    //      WH_KEYBOARD_LL hooks that aggressively drop synthetic injected events
+    //      (like our Backspaces) if they sit in front of us.
+    //   2. Java apps (jp2launcher / javaw / java): commonly embed jnativehook for
+    //      global hotkeys. JVM callback bridge + GC pauses regularly exceed
+    //      Windows' 300ms LowLevelHooksTimeout → Windows drops the hook chain.
+    //      Keeping VKey on top gives us first crack at each event. For
+    //      Vietnamese-eaten keys VKey returns 1 without CallNextHookEx, so a
+    //      downstream stall is irrelevant. For pass-through keys (English
+    //      mode, modifier keys) we still call CallNextHookEx, so a slow
+    //      downstream hook still blocks our callback — partial protection
+    //      only; HookSelfHealer catches the residual case.
+    // Doing this conditionally avoids unnecessary unhook/rehook overhead for
+    // normal apps. We must do this even if the PID hasn't changed — WebView2
+    // creates child windows that trigger focus events AFTER initial hook setup,
+    // and jnativehook may re-arm itself during a JVM session.
+    //
+    // Wave 3 PR 3.2 — gate moved into HookEngine's onReinstall_ lambda
+    // (it checks lifecycle_.ThreadId() and forwards to PostReinstallHooks).
+    // FocusOwner stays decoupled from HookLifecycle's runtime state.
+    if (onReinstall_ && (cls.localElectronApp || cls.isBrowser || cls.isJavaApp)) {
+        const WPARAM reason = cls.isJavaApp ? REINSTALL_REASON_JAVA : REINSTALL_REASON_CHROMIUM;
+        onReinstall_(reason);
+    }
+
+    return cls;
+}
+
+}  // namespace NextKey
